requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
python-dotenv==1.0.0

# NVIDIA NIM SDK
openai==1.6.1  # For NVIDIA NIM API compatibility
langchain==0.1.0
langchain-nvidia-ai-endpoints==0.0.11

# HTML/Web parsing
beautifulsoup4==4.12.2
lxml==4.9.3
html5lib==1.1
cssselect==1.2.0

# Image processing for screenshots
pillow==10.1.0
opencv-python==4.8.1.78
numpy==1.24.3

# Accessibility analysis
axe-selenium-python==2.1.6
wcag-contrast-ratio==0.9

# RAG components
chromadb==0.4.18
sentence-transformers==2.2.2

# API and async
aiohttp==3.9.0
httpx==0.25.2
websockets==12.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Utilities
colorama==0.4.6
tqdm==4.66.1